+++ libkbfs/journal_server.go
@@ -200,55 +200,3 @@ func (j *JournalServer) blockServer() journalBlockServer {
 func (j *JournalServer) mdOps() journalMDOps {
 	return journalMDOps{j.delegateMDOps, j}
 }
-
-type journalBlockServer struct {
-	jServer *JournalServer
-	BlockServer
-}
-
-var _ BlockServer = journalBlockServer{}
-
-func (j journalBlockServer) Put(
-	ctx context.Context, id BlockID, tlfID TlfID, context BlockContext,
-	buf []byte, serverHalf BlockCryptKeyServerHalf) error {
-	_, ok := j.jServer.getBundle(tlfID)
-	if ok {
-		// TODO: Delegate to bundle's block journal.
-	}
-
-	return j.BlockServer.Put(ctx, id, tlfID, context, buf, serverHalf)
-}
-
-func (j journalBlockServer) AddBlockReference(
-	ctx context.Context, id BlockID, tlfID TlfID,
-	context BlockContext) error {
-	_, ok := j.jServer.getBundle(tlfID)
-	if ok {
-		// TODO: Delegate to bundle's block journal.
-	}
-
-	return j.BlockServer.AddBlockReference(ctx, id, tlfID, context)
-}
-
-func (j journalBlockServer) RemoveBlockReference(
-	ctx context.Context, tlfID TlfID,
-	contexts map[BlockID][]BlockContext) (
-	liveCounts map[BlockID]int, err error) {
-	_, ok := j.jServer.getBundle(tlfID)
-	if ok {
-		// TODO: Delegate to bundle's block journal.
-	}
-
-	return j.BlockServer.RemoveBlockReference(ctx, tlfID, contexts)
-}
-
-func (j journalBlockServer) ArchiveBlockReferences(
-	ctx context.Context, tlfID TlfID,
-	contexts map[BlockID][]BlockContext) error {
-	_, ok := j.jServer.getBundle(tlfID)
-	if ok {
-		// TODO: Delegate to bundle's block journal.
-	}
-
-	return j.BlockServer.ArchiveBlockReferences(ctx, tlfID, contexts)
-}
